requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
aiofiles==23.2.1
Pillow==10.1.0
reportlab==4.0.7
matplotlib==3.8.2
pandas==2.1.4